Rename Editor to ReportEditor
