+++ plugins/web/ui/src/routes/report/ReportEditor.svelte
